move the test over to the main package
