+++ pkg/cache/cache_test.go
@@ -2,6 +2,7 @@ package cache_test
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"io"
 	"net/http"
@@ -1296,6 +1297,119 @@ func TestDeleteNar(t *testing.T) {
 			assert.NoFileExists(t, storePath)
 		})
 	})
+}
+
+// TestDeadlock_NarInfo_Triggers_Nar_Refetch reproduces a deadlock where pulling a NarInfo
+// triggers a Nar fetch (because compression is none), and both waiting on each other
+// if they share the same lock/job key.
+func TestDeadlock_NarInfo_Triggers_Nar_Refetch(t *testing.T) {
+	t.Parallel()
+
+	dir, err := os.MkdirTemp("", "cache-path-")
+	require.NoError(t, err)
+
+	defer os.RemoveAll(dir) // clean up
+
+	dbFile := filepath.Join(dir, "var", "ncps", "db", "db.sqlite")
+	testhelper.CreateMigrateDatabase(t, dbFile)
+
+	db, err := database.Open("sqlite:"+dbFile, nil)
+	require.NoError(t, err)
+
+	localStore, err := local.New(newContext(), dir)
+	require.NoError(t, err)
+
+	c, err := newTestCache(newContext(), cacheName, db, localStore, localStore, localStore, "")
+	require.NoError(t, err)
+
+	// 1. Setup a test server
+	ts := testdata.NewTestServer(t, 1)
+	defer ts.Close()
+
+	// CRITICAL: We must ensure NarInfoHash == NarHash to cause the collision in upstreamJobs map.
+	// The deadlock happens because pullNarInfo starts a job with key=hash, and then prePullNar
+	// tries to start a job with key=hash (derived from URL).
+
+	// NarInfoHash is 32 chars.
+	// narURL.Hash comes from URL.
+	// We want narURL.Hash == NarInfoHash.
+	collisionHash := "11111111111111111111111111111111"
+
+	entry := testdata.Entry{
+		NarInfoHash:    collisionHash,
+		NarHash:        collisionHash,
+		NarCompression: "none",
+		NarInfoText: `StorePath: /nix/store/11111111111111111111111111111111-test-1.0
+URL: nar/11111111111111111111111111111111.nar
+Compression: none
+FileHash: sha256:1111111111111111111111111111111111111111111111111111
+FileSize: 123
+NarHash: sha256:1111111111111111111111111111111111111111111111111111
+NarSize: 123
+References: aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-dummy
+Deriver: dddddddddddddddddddddddddddddddd-test-1.0.drv
+Sig: cache.nixos.org-1:MadTCU1OSFCGUw4aqCKpLCZJpqBc7AbLvO7wgdlls0eq1DwaSnF/82SZE+wJGEiwlHbnZR+14daSaec0W3XoBQ==
+`,
+		NarText: "content-of-the-nar",
+	}
+	ts.AddEntry(entry)
+
+	// Add debug handler to see what's being requested and serve content manually
+	ts.AddMaybeHandler(func(w http.ResponseWriter, r *http.Request) bool {
+		if r.URL.Path == "/"+collisionHash+".narinfo" {
+			_, _ = w.Write([]byte(entry.NarInfoText))
+
+			return true
+		}
+
+		if r.URL.Path == "/nar/"+collisionHash+".nar" {
+			_, _ = w.Write([]byte(entry.NarText))
+
+			return true
+		}
+
+		return false // Let the real handler process other things
+	})
+
+	uc, err := upstream.New(newContext(), testhelper.MustParseURL(t, ts.URL), nil)
+	require.NoError(t, err)
+
+	c.AddUpstreamCaches(newContext(), uc)
+
+	// Wait for health check
+	select {
+	case <-c.GetHealthChecker().Trigger():
+	case <-time.After(5 * time.Second):
+		t.Fatal("timeout waiting for upstream health check")
+	}
+
+	// 2. Trigger the download
+	// We use a timeout to detect the deadlock
+	ctx, cancel := context.WithTimeout(newContext(), 5*time.Second)
+	defer cancel()
+
+	done := make(chan struct{})
+
+	var narInfo *narinfo.NarInfo
+
+	go func() {
+		defer close(done)
+
+		narInfo, err = c.GetNarInfo(ctx, entry.NarInfoHash)
+	}()
+
+	select {
+	case <-done:
+	case <-ctx.Done():
+		if errors.Is(ctx.Err(), context.DeadlineExceeded) {
+			t.Fatal("Deadlock detected! GetNarInfo timed out.")
+		}
+	case <-time.After(10 * time.Second):
+		t.Fatal("Timeout waiting for GetNarInfo to complete")
+	}
+
+	require.NoError(t, err)
+	assert.NotNil(t, narInfo)
 }
 
 func newContext() context.Context {
